extracted price-related code
